+++ bin/forrest.php
@@ -17,7 +17,7 @@ $application->setVersion(FORREST_VERSION);
 # Command Commands
 $application->add(new \Startwind\Forrest\CliCommand\Command\ListCommand());
 $application->add(new \Startwind\Forrest\CliCommand\Command\RunCommand());
-$application->add(new \Startwind\Forrest\CliCommand\Command\ShowCommand());
+$application->add(new \Startwind\Forrest\CliCommand\Command\ExplainCommand());
 $application->add(new \Startwind\Forrest\CliCommand\Command\HistoryCommand());
 
 # Repository